+++ dp/_dp_array.py
@@ -12,14 +12,11 @@ class DPArray:
     Args:
         shape (array-like): The dimensions of the array.
         array_name (str): Name of the array, this is used by ``dp.Logger`` when
-            the DP algorithm interacts with multiple arrays.
+            the DP algorithm interacts with multiple arrays. The array name is
+            displayed as the figure title when the array is visualized.
         logger (dp.Logger): Logger object that tracks the actions performed on
             this array, including READ, WRITE, and HIGHLIGHT. This object is
             used to reproduce frame-by-frame animation of the DP algorithm.
-        description_string (string): TODO
-        row_labels ():
-        column_labels ():
-        colors ():
         dtype (str or data-type): Data type of the DPArray. We only support
             ``"f"`` / ``np.float32`` and ``"d"`` / ``np.float64``.
 
@@ -34,10 +31,6 @@ class DPArray:
         array_name="dp_array",
         *,
         logger=None,
-        description_string=None,
-        row_labels=None,
-        column_labels=None,
-        colors=None,
         dtype=np.float64,
     ):
         """Initializes the DPArray."""
@@ -50,10 +43,6 @@ class DPArray:
         self._logger.add_array(array_name, shape)
 
         self._array_name = array_name
-        self._description_string = description_string
-        self._row_labels = row_labels
-        self._column_labels = column_labels
-        self._colors = colors
 
     @staticmethod
     def _parse_dtype(dtype):
@@ -344,6 +333,11 @@ class DPArray:
     def arr(self):
         """Returns the np.ndarray that contains the DP array."""
         return np.array(self._arr, copy=True)
+
+    @property
+    def shape(self):
+        """Returns the shape of the DP array."""
+        return self._arr.shape
 
     @property
     def occupied_arr(self):